+++ src/index.ts
@@ -148,6 +148,7 @@ export {
 } from './lint.js'
 export { tabNormalize } from './tab-normalize.js'
 export { details } from './details.js'
+export { semanticSpan } from './semantic-span.js'
 export { listTable } from './list-table.js'
 export { glossary } from './glossary.js'
 export { headingNumbers, type HeadingNumbersOptions } from './heading-numbers.js'